+++ source/pattern.cpp
@@ -7,9 +7,9 @@ Pattern::Pattern(){
 }
 
 
-Pattern::Pattern(double quality, std::forward_list<unsigned short int> &points){
+Pattern::Pattern(double quality, std::forward_list<unsigned int> &points){
     this->quality = quality;
-    this->cols = std::vector<unsigned short int>{std::make_move_iterator(std::begin(points)), 
+    this->cols = std::vector<unsigned int>{std::make_move_iterator(std::begin(points)), 
                                   std::make_move_iterator(std::end(points)) };
 }
